backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
gspread==6.1.4
google-auth==2.37.0
google-generativeai==0.8.4
